Refactored it a bit
