+++ nx/utils/terminal/__init__.py
@@ -9,11 +9,11 @@ from io import StringIO
 import contextlib
 import logging
 
-from nx.utils.terminal_src.screen import Screen
-from nx.utils.terminal_src.keyboard import Keyboard
-from nx.utils.terminal_src.python import Python
-from nx.utils.terminal_src.menu import Settings
-from nx.utils.terminal_src.utils import Utils
+from .screen import Screen
+from .keyboard import Keyboard
+from .python import Python
+from .menu import Settings
+from .utils import Utils
 
 @contextlib.contextmanager
 def stdoutIO(stdout=None):
@@ -34,8 +34,6 @@ def stderrIO(stderr=None):
     sys.stderr = old
 
 class Terminal(Screen, Keyboard):
-
-
     def __str__(self):
         return "Terminal for the switch, made by PuffDip"
 
@@ -221,5 +219,3 @@ class Terminal(Screen, Keyboard):
             imgui.end()
             imgui.render()
             self.renderer.render()
-
-        self.renderer.shutdown()
