+++ nx/utils/terminal/screen.py
